requirements_with_auth.txt
# Updated requirements.txt - Add these packages for authentication

# Existing packages
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.14.0
pydantic>=2.0.0

# Google Ads API
google-ads>=22.0.0
google-api-python-client>=2.100.0

# Gemini API
google-generativeai>=0.3.0

# NEW: Authentication & OAuth
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
requests>=2.31.0

# NEW: Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# NEW: Security & Sessions
python-jose[cryptography]>=3.3.0
passlib>=1.7.4

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Optional (for advanced features)
lightgbm>=4.0.0  # ML bidding
keybert>=0.8.0  # Keyword extraction
sentence-transformers>=2.2.0
beautifulsoup4>=4.12.0
